Move API agnostic struct code to common files
Moves API agnostic custom struct encode, decode, and to_json routines to
a comnon set of files that are shared by the DX12 and Vulkan specific
code.
