+++ framework/decode/CMakeLists.txt
@@ -38,6 +38,11 @@ target_sources(gfxrecon_decode
                     ${CMAKE_CURRENT_LIST_DIR}/common_struct_handle_mappers.h
                     ${CMAKE_CURRENT_LIST_DIR}/common_object_info_table.cpp
                     ${CMAKE_CURRENT_LIST_DIR}/copy_shaders.h
+                    ${CMAKE_CURRENT_LIST_DIR}/custom_common_struct_decoders.h
+                    ${CMAKE_CURRENT_LIST_DIR}/custom_common_struct_decoders.cpp
+                    ${CMAKE_CURRENT_LIST_DIR}/custom_common_struct_decoders_forward.h
+                    ${CMAKE_CURRENT_LIST_DIR}/custom_common_struct_to_json.h
+                    ${CMAKE_CURRENT_LIST_DIR}/custom_common_struct_to_json.cpp
                     $<$<BOOL:${OPENXR_SUPPORT_ENABLED}>:${CMAKE_CURRENT_LIST_DIR}/custom_openxr_struct_decoders.h>
                     $<$<BOOL:${OPENXR_SUPPORT_ENABLED}>:${CMAKE_CURRENT_LIST_DIR}/custom_openxr_struct_decoders.cpp>
                     $<$<BOOL:${OPENXR_SUPPORT_ENABLED}>:${CMAKE_CURRENT_LIST_DIR}/custom_openxr_struct_decoders_forward.h>
