+++ framework/decode/custom_common_struct_to_json.h
@@ -1,0 +1,42 @@
+/*
+** Copyright (c) 2022-2023 LunarG, Inc.
+** Copyright (c) 2025 Qualcomm Technologies, Inc. and/or its subsidiaries.
+**
+** Permission is hereby granted, free of charge, to any person obtaining a
+** copy of this software and associated documentation files (the "Software"),
+** to deal in the Software without restriction, including without limitation
+** the rights to use, copy, modify, merge, publish, distribute, sublicense,
+** and/or sell copies of the Software, and to permit persons to whom the
+** Software is furnished to do so, subject to the following conditions:
+**
+** The above copyright notice and this permission notice shall be included in
+** all copies or substantial portions of the Software.
+**
+** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+** IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+** FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+** AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+** LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+** FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+** DEALINGS IN THE SOFTWARE.
+*/
+#ifndef GFXRECON_DECODE_CUSTOM_COMMON_STRUCT_TO_JSON_H
+#define GFXRECON_DECODE_CUSTOM_COMMON_STRUCT_TO_JSON_H
+
+#include "decode/custom_common_struct_decoders.h"
+#include "util/defines.h"
+#include "util/json_util.h"
+
+#include "nlohmann/json.hpp"
+
+GFXRECON_BEGIN_NAMESPACE(gfxrecon)
+GFXRECON_BEGIN_NAMESPACE(decode)
+
+void FieldToJson(nlohmann::ordered_json&            jdata,
+                 const Decoded_SECURITY_ATTRIBUTES* data,
+                 const util::JsonOptions&           options = util::JsonOptions());
+
+GFXRECON_END_NAMESPACE(decode)
+GFXRECON_END_NAMESPACE(gfxrecon)
+
+#endif // GFXRECON_DECODE_CUSTOM_COMMON_STRUCT_TO_JSON_H
