+++ framework/encode/custom_vulkan_struct_encoders.cpp
@@ -211,75 +211,6 @@ void EncodeStruct(ParameterEncoder* encoder, const VkAccelerationStructureGeomet
     encoder->EncodeFlagsValue(value.flags);
 }
 
-// The WIN32 SID structure has a variable size, so will be encoded as an array of bytes instead of a struct.
-static void pack_sid_struct(const SID* sid, std::vector<uint8_t>* buffer)
-{
-    assert((sid != nullptr) && (buffer != nullptr));
-
-    const uint8_t* sub_authority      = reinterpret_cast<const uint8_t*>(sid->SubAuthority);
-    size_t         sub_authority_size = sid->SubAuthorityCount * sizeof(uint32_t);
-
-    // sizeof(SID) already includes the size of one of the SidAuthority elements,
-    // so we can subtract 4 bytes from sid_authority_size.
-    buffer->reserve(sizeof(SID) + (sub_authority_size - sizeof(uint32_t)));
-
-    buffer->push_back(sid->Revision);
-    buffer->push_back(sid->SubAuthorityCount);
-    buffer->insert(buffer->end(), sid->IdentifierAuthority.Value, sid->IdentifierAuthority.Value + 6);
-    buffer->insert(buffer->end(), sub_authority, sub_authority + sub_authority_size);
-}
-
-void EncodeStruct(ParameterEncoder* encoder, const ACL& value)
-{
-    encoder->EncodeUInt8Value(value.AclRevision);
-    encoder->EncodeUInt8Value(value.Sbz1);
-    encoder->EncodeUInt16Value(value.AclSize);
-    encoder->EncodeUInt16Value(value.AceCount);
-    encoder->EncodeUInt16Value(value.Sbz2);
-}
-
-void EncodeStruct(ParameterEncoder* encoder, const SECURITY_DESCRIPTOR& value)
-{
-    encoder->EncodeUInt8Value(value.Revision);
-    encoder->EncodeUInt8Value(value.Sbz1);
-    encoder->EncodeUInt16Value(value.Control);
-
-    // The SID structure has a variable size, so will be packed into an array of bytes.
-    std::vector<uint8_t> buffer;
-
-    if (value.Owner != nullptr)
-    {
-        pack_sid_struct(reinterpret_cast<SID*>(value.Owner), &buffer);
-        encoder->EncodeUInt8Array(buffer.data(), buffer.size());
-    }
-    else
-    {
-        encoder->EncodeUInt8Array(nullptr, 0);
-    }
-
-    buffer.clear();
-
-    if (value.Group != nullptr)
-    {
-        pack_sid_struct(reinterpret_cast<SID*>(value.Group), &buffer);
-        encoder->EncodeUInt8Array(buffer.data(), buffer.size());
-    }
-    else
-    {
-        encoder->EncodeUInt8Array(nullptr, 0);
-    }
-
-    EncodeStructPtr(encoder, value.Sacl);
-    EncodeStructPtr(encoder, value.Dacl);
-}
-
-void EncodeStruct(ParameterEncoder* encoder, const SECURITY_ATTRIBUTES& value)
-{
-    encoder->EncodeUInt32Value(value.nLength);
-    EncodeStructPtr(encoder, reinterpret_cast<SECURITY_DESCRIPTOR*>(value.lpSecurityDescriptor));
-    encoder->EncodeInt32Value(value.bInheritHandle);
-}
-
 void EncodeStruct(ParameterEncoder* encoder, const VkPushDescriptorSetWithTemplateInfoKHR& value)
 {
     encoder->EncodeEnumValue(value.sType);
